feat: Updated GraphQL form UI (#36728)
## Description

This PR has following new features and fixes
1. Feat: Update GraphQL to use form control dynamic text field instead
of custom implementation.
2. Feat: Added full width support for section component in action form.
3. Fix: Rest API headers not scrollable if there are more number of
items.

EE PR: https://github.com/appsmithorg/appsmith-ee/pull/5297

Fixes #35494
Fixes #36410
Fixes #36499


## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
