+++ app/client/src/pages/Editor/APIEditor/CommonEditorForm.tsx
@@ -111,6 +111,7 @@ const Wrapper = styled.div`
   flex-direction: row;
   height: 100%;
   position: relative;
+  overflow: hidden;
 `;
 
 const MainContainer = styled.div`
