+++ app/client/src/pages/Editor/ActionForm/Section/index.tsx
@@ -5,11 +5,13 @@ import styles from "./styles.module.css";
 interface SectionProps extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
   isStandalone?: boolean;
+  isFullWidth?: boolean;
 }
 
 const Section: React.FC<SectionProps> = ({
   children,
   className,
+  isFullWidth = false,
   isStandalone = false,
   ...props
 }) => {
@@ -18,6 +20,7 @@ const Section: React.FC<SectionProps> = ({
   return (
     <div
       className={classNames}
+      data-fullwidth={isFullWidth.toString()}
       data-standalone={isStandalone.toString()}
       {...props}
     >
