+++ app/client/src/pages/Editor/ActionForm/Section/styles.module.css
@@ -13,4 +13,8 @@
   &[data-standalone="false"]:not(:last-child) {
     border-bottom: 1px solid var(--ads-v2-color-border);
   }
+
+  &[data-fullwidth="true"] {
+    max-width: none;
+  }
 }
